reuse more code and cleanup makefiles.
