+++ examples/disk_compressor/disk_compressor.pro
@@ -1,4 +1,5 @@
 include(../../QArchive.pri)
+DEFINES += BUILD_EXAMPLES
 TEMPLATE = app
 TARGET = compressor
 SOURCES += main.cc
