+++ examples/disk_extractor_with_QIODevice/disk_extractor_with_QIODevice.pro
@@ -1,4 +1,5 @@
 include(../../QArchive.pri)
+DEFINES += BUILD_EXAMPLES
 TEMPLATE = app
 TARGET = extractor
 SOURCES += main.cc
