+++ examples/examples.pro
@@ -3,4 +3,5 @@ CONFIG += ordered
 
 SUBDIRS = disk_extractor \
           disk_compressor \
-	  disk_extractor_with_QIODevice
+	  disk_extractor_with_QIODevice \
+	  memory_extractor
