+++ examples/memory_extractor/memory_extractor.pro
@@ -1,4 +1,5 @@
 include(../../QArchive.pri)
 TEMPLATE = app
+DEFINES += BUILD_EXAMPLES
 TARGET = extractor
 SOURCES += main.cc
